renamed jitsi-bot
