+++ src/commands/joke.cmd.ts
@@ -1,5 +1,5 @@
 import https from 'https';
-import JitsiBot from '../bot';
+import JitsiBot from '../jitsi-bot';
 import { IJimCommand } from '../command.service';
 import config from '../config';
 import { IIncomingMessage } from '../models/jitsi.interface';
